+++ Sources/Libraries/imgui/widgets/label.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2012-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2012-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 //
 // This library is distributed under the MIT License. See notice at the end
 // of this file.
@@ -22,7 +22,7 @@ namespace imgui
 	{
 		core::checked_begin();
 		float h = 30.0f;
-		detail::font->resource([&h](auto& f) { h = f.face.height; });
+		detail::font->resource([&h](auto& f) { h = static_cast<float>(f.face.height); });
 		Control ctrl(controls::LABEL, Rect(pos.x, pos.y, 100.0f, h));
 		unigui::draw::print(text, ctrl.rect, detail::font, unigui::color(1.0f, 1.0f, 1.0f, 1.0f), true);
 		core::checked_end();
@@ -47,7 +47,7 @@ namespace imgui
 
 
 
-// Copyright (C) 2012-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2012-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated 
 // documentation files (the "Software"), to deal in the Software without restriction, including without limitation
