+++ Sources/Libraries/uniguidraw/drawers.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2012-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2012-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 //
 // This library is distributed under the MIT License. See notice at the end
 // of this file.
@@ -553,8 +553,8 @@ namespace unigui
 
 		void textureSize(float& w, float& h, multiapi::Texture* texture)
 		{
-			w = texture->sizeX;
-			h = texture->sizeY;
+			w = (float) texture->sizeX;
+			h = (float) texture->sizeY;
 		}
 
 		void textureSize(float& w, float& h, Resources::Texture* texture)
@@ -605,7 +605,7 @@ namespace unigui
 
 
 
-// Copyright (C) 2012-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2012-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated 
 // documentation files (the "Software"), to deal in the Software without restriction, including without limitation
